# backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LocalMind API",
        "description": "OpenAPI documentation for the LocalMind backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Docker-served LocalMind backend"
        }
    ],
    "paths": {
...
        "/api/questions/{question}/responses": {
            "post": {
                "tags": [
                    "Responses"
                ],
                "summary": "Create a response for a question",
                "operationId": "57225c9b32e57bc210690744f2c00135",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResponseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Response created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseMutationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/responses/{response}": {
            "put": {
                "tags": [
                    "Responses"
                ],
                "summary": "Update a response",
                "operationId": "7683bb4709fa02456a33cceea7d24e64",
                "parameters": [
                    {
                        "name": "response",
                        "in": "path",
                        "description": "Response id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResponseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseMutationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Responses"
                ],
                "summary": "Delete a response",
                "operationId": "4f0396624a62d4167b1b6ff0613145c0",
                "parameters": [
                    {
                        "name": "response",
                        "in": "path",
                        "description": "Response id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthorSummary": {
                "required": [
                    "id",
                    "name",
                    "initial"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lahcen"
                    },
                    "initial": {
                        "type": "string",
                        "example": "L"
                    }
                },
                "type": "object"
            },
...
            "ResponseResource": {
                "required": [
                    "id",
                    "content",
                    "question_id",
                    "created_at",
                    "author",
                    "is_owner"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "content": {
                        "type": "string",
                        "example": "You can find a pharmacy near the train station."
                    },
                    "question_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2 minutes ago"
                    },
                    "author": {
                        "$ref": "#/components/schemas/AuthorSummary"
                    },
                    "is_owner": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "ResponseMutationResponse": {
                "required": [
                    "message",
                    "data"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Intelligence contribution recorded."
                    },
                    "data": {
                        "$ref": "#/components/schemas/ResponseResource"
                    }
                },
                "type": "object"
            },
...
            "MessageResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operation completed successfully."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
...
            "ResponseRequest": {
                "required": [
                    "content"
                ],
                "properties": {
                    "content": {
                        "type": "string",
                        "example": "There is one next to the central market.",
                        "maxLength": 2000,
                        "minLength": 5
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the bearer token returned by the register or login endpoint.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}